Add nosetests command to setup script
Instead of running `nosetests` directly which does not install the
required packages for testing, the 'nosetests' command is added as it is
described in Nose documentation. It is also added to `setup_requires`
list to make sure that nose is installed automatically before running
test cases.

Consequently, there is no need to have test rule in the Makefile. Since
this commit, Makefile is only for convenience use and all packaging and
distributing actions are done by setup script.