Update setup script
Testing, packaging, distributing should be done by setup script. The
Makefile is just for the matter of conveniences and a wrapper on setup
script commands (`build`, `sdist`, ...) except for compiling the sample
proto file in the test package.

The package dependencies will be installed by setup script, so there is
no need to install the requirements in the `init` rule in Makefile.

For security reasons, it is recommended to use 'twine' instead of
`register`, and `upload` commands in setup script. So, the distribution
rules in Makefile is updated to use 'twine'.

Package metadata is also updated to identify the license file.

Finally, the `keywords` parameter passed to setup function call should
be a string of keywords separated by space; not list.